# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/app/settings/theme:
    get:
      operationId: SettingsController_getTheme
      parameters: []
      responses:
        '200':
          description: Configurazione tema corrente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeConfigDto'
      security:
        - access-token: []
      summary: Tema globale dell'installazione (default di fabbrica se mai salvato)
      tags:
        - Settings
    put:
      operationId: SettingsController_updateTheme
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThemeConfigDto'
      responses:
        '200':
          description: Tema salvato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeConfigDto'
        '400':
          description: Payload non valido (hex, palette o versione)
        '403':
          description: Ruolo non SuperAdmin
      security:
        - access-token: []
      summary: Salva il tema globale (SuperAdmin only, registrato su audit log)
      tags:
        - Settings
...
  /api/v1/app/files:
    post:
      operationId: FilesController_upload
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadFileDto'
      responses:
        '201':
          description: File caricato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadataDto'
        '413':
          description: File più grande del limite configurato
      security:
        - access-token: []
      summary: Carica un documento (multipart/form-data, campo "file")
      tags:
        - Files
  /api/v1/app/files/{guid}:
    get:
      operationId: FilesController_download
      parameters:
        - name: guid
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Contenuto del file
        '404':
          description: File non trovato o eliminato
      security:
        - access-token: []
      summary: Scarica il contenuto di un file (streaming)
      tags:
        - Files
    delete:
      operationId: FilesController_delete
      parameters:
        - name: guid
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: File eliminato
        '403':
          description: Non sei l'autore del file e non hai un ruolo Admin/superiore
        '404':
          description: File non trovato o già eliminato
      security:
        - access-token: []
      summary: >-
        Elimina un file (soft-delete, blob fisico non rimosso subito — vedi
        ADR-8)
      tags:
        - Files
  /api/v1/app/notifications:
    get:
      operationId: NotificationsController_findAll
      parameters:
        - name: p
          required: false
          in: query
          description: Pagina (default 1)
          schema:
            type: string
        - name: i
          required: false
          in: query
          description: Elementi per pagina (default 20)
          schema:
            type: string
        - name: unreadOnly
          required: false
          in: query
          description: Se "true", restituisce solo le notifiche non lette
          schema:
            type: string
      responses:
        '200':
          description: Lista notifiche paginata
      security:
        - access-token: []
      summary: Lista paginata delle notifiche del chiamante, più recenti prima
      tags:
        - Notifications
  /api/v1/app/notifications/unread-count:
    get:
      operationId: NotificationsController_unreadCount
      parameters: []
      responses:
        '200':
          description: Conteggio non lette
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCountDto'
      security:
        - access-token: []
      summary: Numero di notifiche non lette del chiamante
      tags:
        - Notifications
  /api/v1/app/notifications/{guid}/read:
    patch:
      operationId: NotificationsController_markRead
      parameters:
        - name: guid
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Notifica aggiornata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationDto'
        '404':
          description: Notifica non trovata o non del chiamante
      security:
        - access-token: []
      summary: Segna una notifica come letta (solo se del chiamante)
      tags:
        - Notifications
  /api/v1/app/notifications/read-all:
    patch:
      operationId: NotificationsController_markAllRead
      parameters: []
      responses:
        '200':
          description: Numero di notifiche aggiornate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkAllReadDto'
      security:
        - access-token: []
      summary: Segna tutte le notifiche del chiamante come lette
      tags:
        - Notifications
info:
  title: Starter Kit API
  description: >-
    API REST del boilerplate aziendale (auth, RBAC, MFA, audit log, gestione
    utenti).
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ThemeColorsDto:
      type: object
      properties:
        primary:
          type: string
          description: Colore primario
          example: '#228be6'
        secondary:
          type: string
          description: Colore secondario
          example: '#868e96'
        accent:
          type: string
          description: Colore accento
          example: '#be4bdb'
        success:
          type: string
          description: Colore successo
          example: '#40c057'
        warning:
          type: string
          description: Colore avviso
          example: '#fab005'
        alert:
          type: string
          description: Colore allerta
          example: '#f76707'
        error:
          type: string
          description: Colore errore
          example: '#fa5252'
        danger:
          type: string
          description: Colore pericolo
          example: '#c92a2a'
        info:
          type: string
          description: Colore informativo
          example: '#15aabf'
      required:
        - primary
        - secondary
        - accent
        - success
        - warning
        - alert
        - error
        - danger
        - info
    ThemePrimaryShadeDto:
      type: object
      properties:
        light:
          type: number
          description: Shade filled nello scheme chiaro (0–9)
          example: 8
        dark:
          type: number
          description: Shade filled nello scheme scuro (0–9)
          example: 5
      required:
        - light
        - dark
    ThemeGradientDto:
      type: object
      properties:
        from:
          type: string
          description: Colore di partenza
          example: '#228be6'
        to:
          type: string
          description: Colore di arrivo
          example: '#15aabf'
        deg:
          type: number
          description: Angolo in gradi (0–360)
          example: 45
      required:
        - from
        - to
        - deg
    ThemeFontSizesDto:
      type: object
      properties:
        xs:
          type: number
          description: Dimensione testo xs
          example: 12
        sm:
          type: number
          description: Dimensione testo sm
          example: 14
        md:
          type: number
          description: Dimensione testo md
          example: 16
        lg:
          type: number
          description: Dimensione testo lg
          example: 18
        xl:
          type: number
          description: Dimensione testo xl
          example: 20
      required:
        - xs
        - sm
        - md
        - lg
        - xl
    ThemeLineHeightsDto:
      type: object
      properties:
        xs:
          type: number
          description: Interlinea xs
          example: 1.4
        sm:
          type: number
          description: Interlinea sm
          example: 1.45
        md:
          type: number
          description: Interlinea md
          example: 1.55
        lg:
          type: number
          description: Interlinea lg
          example: 1.6
        xl:
          type: number
          description: Interlinea xl
          example: 1.65
      required:
        - xs
        - sm
        - md
        - lg
        - xl
...
    ThemeHeadingsDto:
      type: object
      properties:
        fontFamily:
          type: string
          description: Font dei titoli (ID whitelisted)
          enum:
            - inter
            - system
            - humanist
            - geometric
            - rounded
            - serif
            - slab
          example: inter
        fontWeight:
          type: string
          description: Peso dei titoli
          enum:
            - '300'
            - '400'
            - '500'
            - '600'
            - '700'
            - '800'
            - '900'
          example: '700'
        fontSizeUnit:
          type: string
          description: Unità CSS condivisa dalla dimensione di ogni livello h1–h6 (v7)
          enum:
            - px
            - em
            - rem
            - '%'
          example: px
        sizes:
          description: Dimensioni h1–h6
          allOf:
            - $ref: '#/components/schemas/ThemeHeadingSizesDto'
      required:
        - fontFamily
        - fontWeight
        - fontSizeUnit
        - sizes
    ThemeTypographyDto:
      type: object
      properties:
        fontFamily:
          type: string
          description: Font del testo (ID whitelisted)
          enum:
            - inter
            - system
            - humanist
            - geometric
            - rounded
            - serif
            - slab
          example: inter
        fontFamilyMonospace:
          type: string
          description: Font monospace (ID whitelisted)
          enum:
            - system-mono
            - courier
          example: system-mono
        fontSizes:
          description: Dimensioni testo xs–xl
          allOf:
            - $ref: '#/components/schemas/ThemeFontSizesDto'
        fontSizeUnit:
          type: string
          description: Unità CSS di fontSizes (v7)
          enum:
            - px
            - em
            - rem
            - '%'
          example: px
        lineHeights:
          description: Interlinee xs–xl
          allOf:
            - $ref: '#/components/schemas/ThemeLineHeightsDto'
        headings:
          description: Configurazione titoli
          allOf:
            - $ref: '#/components/schemas/ThemeHeadingsDto'
      required:
        - fontFamily
        - fontFamilyMonospace
        - fontSizes
        - fontSizeUnit
        - lineHeights
        - headings
    ThemeSpacingDto:
      type: object
      properties:
        xs:
          type: number
          description: Spaziatura xs
          example: 10
        sm:
          type: number
          description: Spaziatura sm
          example: 12
        md:
          type: number
          description: Spaziatura md
          example: 16
        lg:
          type: number
          description: Spaziatura lg
          example: 20
        xl:
          type: number
          description: Spaziatura xl
          example: 32
      required:
        - xs
        - sm
        - md
        - lg
        - xl
    ThemeRadiusScaleDto:
      type: object
      properties:
        xs:
          type: number
          description: Radius xs
          example: 2
        sm:
          type: number
          description: Radius sm
          example: 4
        md:
          type: number
          description: Radius md
          example: 8
        lg:
          type: number
          description: Radius lg
          example: 16
        xl:
          type: number
          description: Radius xl
          example: 32
      required:
        - xs
        - sm
        - md
        - lg
        - xl
    ThemeShadowSpecDto:
      type: object
      properties:
        'y':
          type: number
          description: Offset verticale
          example: 1
        blur:
          type: number
          description: Sfocatura
          example: 3
        spread:
          type: number
          description: Espansione
          example: 0
        opacity:
          type: number
          description: Opacità del nero (0–1)
          example: 0.05
      required:
        - 'y'
        - blur
        - spread
        - opacity
    ThemeShadowsDto:
      type: object
      properties:
        xs:
          description: Ombra xs
          allOf:
            - $ref: '#/components/schemas/ThemeShadowSpecDto'
        sm:
          description: Ombra sm
          allOf:
            - $ref: '#/components/schemas/ThemeShadowSpecDto'
        md:
          description: Ombra md
          allOf:
            - $ref: '#/components/schemas/ThemeShadowSpecDto'
        lg:
          description: Ombra lg
          allOf:
            - $ref: '#/components/schemas/ThemeShadowSpecDto'
        xl:
          description: Ombra xl
          allOf:
            - $ref: '#/components/schemas/ThemeShadowSpecDto'
      required:
        - xs
        - sm
        - md
        - lg
        - xl
    ThemeButtonDefaultsDto:
      type: object
      properties:
        variant:
          type: string
          description: Variant di default
          enum:
            - unset
            - filled
            - light
            - outline
            - subtle
            - default
            - gradient
          example: unset
        size:
          type: string
          description: Size di default
          enum:
            - unset
            - xs
            - sm
            - md
            - lg
            - xl
          example: unset
        radius:
          type: string
          description: Radius di default
          enum:
            - unset
            - xs
            - sm
            - md
            - lg
            - xl
          example: unset
      required:
        - variant
        - size
        - radius
...
    ThemeComponentsDto:
      type: object
      properties:
        button:
          description: Default Button
          allOf:
            - $ref: '#/components/schemas/ThemeButtonDefaultsDto'
        actionIcon:
          description: Default ActionIcon
          allOf:
            - $ref: '#/components/schemas/ThemeActionIconDefaultsDto'
        badge:
          description: Default Badge
          allOf:
            - $ref: '#/components/schemas/ThemeBadgeDefaultsDto'
        input:
          description: Default campi input
          allOf:
            - $ref: '#/components/schemas/ThemeInputDefaultsDto'
        card:
          description: Default Paper/Card
          allOf:
            - $ref: '#/components/schemas/ThemeCardDefaultsDto'
        modal:
          description: Default Modal/Drawer
          allOf:
            - $ref: '#/components/schemas/ThemeModalDefaultsDto'
        table:
          description: Default Table
          allOf:
            - $ref: '#/components/schemas/ThemeTableDefaultsDto'
        tooltip:
          description: Default Tooltip
          allOf:
            - $ref: '#/components/schemas/ThemeTooltipDefaultsDto'
        loader:
          description: Default Loader
          allOf:
            - $ref: '#/components/schemas/ThemeLoaderDefaultsDto'
      required:
        - button
        - actionIcon
        - badge
        - input
        - card
        - modal
        - table
        - tooltip
        - loader
    ThemeSchemeTokensDto:
      type: object
      properties:
        pageBg:
          type: string
          description: Sfondo applicativo
          example: '#f8f9fa'
        cardBg:
          type: string
          description: Sfondo card / superfici contenuto
          example: '#ffffff'
        cardBorder:
          type: string
          description: Bordo card
          example: '#ffffff'
        textPrimary:
          type: string
          description: Testo principale
          example: '#000000'
        textSecondary:
          type: string
          description: Testo secondario/dimmed
          example: '#868e96'
        headingH1:
          type: string
          description: Colore del titolo H1
          example: '#000000'
        headingH2:
          type: string
          description: Colore del titolo H2
          example: '#000000'
        headingH3:
          type: string
          description: Colore del titolo H3
          example: '#000000'
        headingH4:
          type: string
          description: Colore del titolo H4
          example: '#000000'
        headingH5:
          type: string
          description: Colore del titolo H5
          example: '#000000'
        headingH6:
          type: string
          description: Colore del titolo H6
          example: '#000000'
        navbarBg:
          type: string
          description: Sfondo sidebar
          example: '#242424'
        navbarText:
          type: string
          description: Testo voci navbar
          example: '#b8b8b8'
        navbarHoverBg:
          type: string
          description: Sfondo hover voce navbar
          example: '#2e2e2e'
        navbarActiveBg:
          type: string
          description: Sfondo voce navbar attiva
          example: '#1971c2'
        navbarActiveText:
          type: string
          description: Testo voce navbar attiva
          example: '#ffffff'
        navbarBorder:
          type: string
          description: Bordi interni sidebar
          example: '#3b3b3b'
      required:
        - pageBg
        - cardBg
        - cardBorder
        - textPrimary
        - textSecondary
        - headingH1
        - headingH2
        - headingH3
        - headingH4
        - headingH5
        - headingH6
        - navbarBg
        - navbarText
        - navbarHoverBg
        - navbarActiveBg
        - navbarActiveText
        - navbarBorder
    ThemeConfigDto:
      type: object
      properties:
        version:
          type: number
          description: Versione del contratto ThemeConfig
          example: 7
          enum:
            - 7
        navbarWidth:
          type: number
          description: Larghezza sidebar espansa
          example: 210
        navbarWidthUnit:
          type: string
          description: Unità CSS di navbarWidth (v7)
          enum:
            - px
            - em
            - rem
            - '%'
          example: px
        navbarDefaultCollapsed:
          type: boolean
          description: Sidebar chiusa (solo icone) di default al caricamento
          example: false
        navbarEdgeStyle:
          type: string
          description: >-
            Stile del bordo destro della sidebar: bordo sottile o ombra
            proiettata
          enum:
            - border
            - shadow
          example: border
        navbarEdgeShadowIntensity:
          type: number
          description: >-
            Intensità (0–1) dell'ombra del bordo destro sidebar quando
            navbarEdgeStyle è 'shadow'
          example: 0.16
        colors:
          description: I 9 colori semantici del tema
          allOf:
            - $ref: '#/components/schemas/ThemeColorsDto'
        primaryShade:
          description: Shade filled per scheme
          allOf:
            - $ref: '#/components/schemas/ThemePrimaryShadeDto'
        radius:
          type: string
          description: Radius di default dei componenti
          enum:
            - xs
            - sm
            - md
            - lg
            - xl
          example: md
        focusRing:
          type: string
          description: Anello di focus
          enum:
            - auto
            - always
            - never
          example: auto
        cursorType:
          type: string
          description: Cursore sui controlli interattivi
          enum:
            - default
            - pointer
          example: default
        respectReducedMotion:
          type: boolean
          description: Rispetta prefers-reduced-motion
          example: false
        autoContrast:
          type: boolean
          description: Contrasto automatico sui filled
          example: false
        luminanceThreshold:
          type: number
          description: Soglia di luminanza per autoContrast (0–1)
          example: 0.3
        scale:
          type: number
          description: Scala globale interfaccia (0.75–1.5)
          example: 1
        defaultGradient:
          description: Gradiente di default
          allOf:
            - $ref: '#/components/schemas/ThemeGradientDto'
        typography:
          description: Blocco tipografico
          allOf:
            - $ref: '#/components/schemas/ThemeTypographyDto'
        spacing:
          description: Scala di spaziatura xs–xl
          allOf:
            - $ref: '#/components/schemas/ThemeSpacingDto'
        spacingUnit:
          type: string
          description: Unità CSS di spacing (v7)
          enum:
            - px
            - em
            - rem
            - '%'
          example: px
        radiusScale:
          description: Valori dei radius token xs–xl
          allOf:
            - $ref: '#/components/schemas/ThemeRadiusScaleDto'
        radiusScaleUnit:
          type: string
          description: Unità CSS di radiusScale (v7)
          enum:
            - px
            - em
            - rem
            - '%'
          example: px
        shadows:
          description: Ombre xs–xl strutturate
          allOf:
            - $ref: '#/components/schemas/ThemeShadowsDto'
        shadowUnit:
          type: string
          description: Unità CSS di shadows (v7) — solo lunghezze, mai %
          enum:
            - px
            - em
            - rem
          example: px
        components:
          description: Default per-componente
          allOf:
            - $ref: '#/components/schemas/ThemeComponentsDto'
        light:
          description: Token per lo scheme chiaro
          allOf:
            - $ref: '#/components/schemas/ThemeSchemeTokensDto'
        dark:
          description: Token per lo scheme scuro
          allOf:
            - $ref: '#/components/schemas/ThemeSchemeTokensDto'
      required:
        - version
        - navbarWidth
        - navbarWidthUnit
        - navbarDefaultCollapsed
        - navbarEdgeStyle
        - navbarEdgeShadowIntensity
        - colors
        - primaryShade
        - radius
        - focusRing
        - cursorType
        - respectReducedMotion
        - autoContrast
        - luminanceThreshold
        - scale
        - defaultGradient
        - typography
        - spacing
        - spacingUnit
        - radiusScale
        - radiusScaleUnit
        - shadows
        - shadowUnit
        - components
        - light
        - dark
    UploadFileDto:
      type: object
      properties:
        entity:
          type: string
          description: Nome tabella/dominio a cui associare il file
          example: invoice
          maxLength: 100
        entityId:
          type: string
          description: Id/guid dell'entità di dominio da associare
          example: a1b2c3d4e5f6a7b8
          maxLength: 100
    FileMetadataDto:
      type: object
      properties:
        guid:
          type: string
          description: Identificatore pubblico del file, usato nelle URL
          example: a1b2c3d4e5f6a7b8
        originalName:
          type: string
          description: Nome file originale (solo display)
          example: fattura-2026-001.pdf
        mimeType:
          type: string
          description: MIME type dichiarato dal client
          example: application/pdf
        sizeBytes:
          type: number
          description: Dimensione del file in byte
          example: 348213
        entity:
          type: object
          description: Nome tabella/dominio a cui il file è associato, se presente
          example: invoice
          nullable: true
        entityId:
          type: object
          description: Id/guid dell'entità di dominio associata, se presente
          example: a1b2c3d4e5f6a7b8
          nullable: true
        createdAt:
          format: date-time
          type: string
          description: Data di caricamento
          example: '2026-07-23T10:00:00.000Z'
      required:
        - guid
        - originalName
        - mimeType
        - sizeBytes
        - createdAt
    UnreadCountDto:
      type: object
      properties:
        count:
          type: number
          description: Numero di notifiche non lette del chiamante
          example: 3
      required:
        - count
    NotificationDto:
      type: object
      properties:
        guid:
          type: string
          description: Identificatore pubblico della notifica, usato nelle URL
          example: a1b2c3d4e5f6a7b8
        type:
          type: string
          description: Codice tipo libero, definito dal progetto verticale
          example: system.info
        title:
          type: string
          description: Titolo breve
          example: Nuovo documento caricato
        message:
          type: string
          description: Testo del messaggio
          example: Il file "fattura.pdf" è stato caricato con successo.
        link:
          type: object
          description: Percorso frontend su cui portare l'utente al click, se presente
          example: /files
          nullable: true
        isRead:
          type: boolean
          description: Se la notifica è già stata letta
          example: false
        createdAt:
          format: date-time
          type: string
          description: Data di creazione
          example: '2026-07-23T10:00:00.000Z'
      required:
        - guid
        - type
        - title
        - message
        - isRead
        - createdAt
    MarkAllReadDto:
      type: object
      properties:
        updated:
          type: number
          description: Numero di notifiche aggiornate
          example: 3
      required:
        - updated